moves scoped_device_id implementations into module code
