+++ core/base/scoped_device_id.cpp
@@ -39,70 +39,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/config.hpp>
 
 
-#include "core/base/scoped_device_id.hpp"
+#include "core/base/noop_scoped_device_id.hpp"
 
 
 namespace gko {
-namespace detail {
-
-
-cuda_scoped_device_id::cuda_scoped_device_id(
-    cuda_scoped_device_id&& other) noexcept
-{
-    *this = std::move(other);
-}
-
-
-cuda_scoped_device_id& cuda_scoped_device_id::operator=(
-    cuda_scoped_device_id&& other) noexcept
-{
-    if (this != &other) {
-        original_device_id_ = std::exchange(other.original_device_id_, 0);
-        need_reset_ = std::exchange(other.need_reset_, false);
-    }
-    return *this;
-}
-
-
-hip_scoped_device_id::hip_scoped_device_id(
-    hip_scoped_device_id&& other) noexcept
-{
-    *this = std::move(other);
-}
-
-
-hip_scoped_device_id& hip_scoped_device_id::operator=(
-    hip_scoped_device_id&& other) noexcept
-{
-    if (this != &other) {
-        original_device_id_ = std::exchange(other.original_device_id_, 0);
-        need_reset_ = std::exchange(other.need_reset_, false);
-    }
-    return *this;
-}
-
-
-}  // namespace detail
-
-
-scoped_device_id::scoped_device_id(const OmpExecutor* exec, int device_id)
-    : scope_(std::make_unique<detail::noop_scoped_device_id>())
-{}
-
-
-scoped_device_id::scoped_device_id(const CudaExecutor* exec, int device_id)
-    : scope_(std::make_unique<detail::cuda_scoped_device_id>(device_id))
-{}
-
-
-scoped_device_id::scoped_device_id(const HipExecutor* exec, int device_id)
-    : scope_(std::make_unique<detail::hip_scoped_device_id>(device_id))
-{}
-
-
-scoped_device_id::scoped_device_id(const DpcppExecutor* exec, int device_id)
-    : scope_(std::make_unique<detail::noop_scoped_device_id>())
-{}
-
-
+namespace detail {}  // namespace detail
 }  // namespace gko
