+++ cuda/base/executor.cpp
@@ -44,12 +44,10 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/base/exception_helpers.hpp>
 
 
-#include "core/base/scoped_device_id.hpp"
-
-
 #include "cuda/base/config.hpp"
 #include "cuda/base/cublas_bindings.hpp"
 #include "cuda/base/cusparse_handle.hpp"
+#include "cuda/base/scoped_device_id.hpp"
 
 
 namespace gko {
