+++ hip/base/executor.hip.cpp
@@ -44,12 +44,10 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/base/exception_helpers.hpp>
 
 
-#include "core/base/scoped_device_id.hpp"
-
-
 #include "hip/base/config.hip.hpp"
 #include "hip/base/hipblas_bindings.hip.hpp"
 #include "hip/base/hipsparse_bindings.hip.hpp"
+#include "hip/base/scoped_device_id.hip.hpp"
 
 
 namespace gko {
