+++ hip/base/scoped_device_id.hip.cpp
@@ -30,11 +30,8 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************<GINKGO LICENSE>*******************************/
 
-#ifndef GKO_HIP_BASE_SCOPED_DEVICE_ID_HIP_HPP_
-#define GKO_HIP_BASE_SCOPED_DEVICE_ID_HIP_HPP_
-
-
 #include <exception>
+#include <utility>
 
 
 #include <hip/hip_runtime.h>
@@ -43,7 +40,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/base/exception_helpers.hpp>
 
 
-#include "core/base/scoped_device_id.hpp"
+#include "hip/base/scoped_device_id.hip.hpp"
 
 
 namespace gko {
@@ -74,8 +71,30 @@ hip_scoped_device_id::~hip_scoped_device_id() noexcept(false)
 }
 
 
+hip_scoped_device_id::hip_scoped_device_id(
+    hip_scoped_device_id&& other) noexcept
+{
+    *this = std::move(other);
+}
+
+
+hip_scoped_device_id& hip_scoped_device_id::operator=(
+    gko::detail::hip_scoped_device_id&& other) noexcept
+{
+    if (this != &other) {
+        original_device_id_ = std::exchange(other.original_device_id_, 0);
+        need_reset_ = std::exchange(other.need_reset_, false);
+    }
+    return *this;
+}
+
+
 }  // namespace detail
+
+
+scoped_device_id::scoped_device_id(const HipExecutor* exec, int device_id)
+    : scope_(std::make_unique<detail::hip_scoped_device_id>(device_id))
+{}
+
+
 }  // namespace gko
-
-
-#endif  // GKO_HIP_BASE_SCOPED_DEVICE_ID_HIP_HPP_
